#120: re-factoring file layout of gem package to be more correctly named
